+++ src/tap.ts
@@ -1,6 +1,4 @@
-export function tap<T>(fn: (value: Readonly<T>) => any): (a: T) => T {
-  return (value: T) => {
-    fn(value);
-    return value;
-  };
+export function tap<T>(value: T, fn: (value: Readonly<T>) => any): T {
+  fn(value);
+  return value;
 }